Create standard_sol2
29 / 29 test cases passed.
Status: Accepted
Runtime: 1 ms
Memory Usage: 39.8 MB

This trick is interesting. Maybe we can call this kind of question "symmetric", because each solution comes in pairs, so we don't have to consider the order of two values in a specific pair. Moreover, hashmap also doesn't have the concept of order. So we can use it to do this kind of symmetric problem. 

I at first put map.put(nums[i],i); in front of the for loop and it doesn't work. Because when two values all the same (such as [3,3]), when I add this to the hashtable, which cannot have duplicate key, the second 3 will replace the first 3, so when target is 6, I cannot get any solution. If I put this in after for loop, then I have already checked values before current value and it is okay to replace them if they are the same.